Changes how we determine verified channels. Adds ui elements for match status.
Needed to change how we access the verified channels. Also needed to
change the getMatch function to send match data on successful query so
the user does not need to wait for a match status change. Needed ui
elements to show the user when their data is being successfully recorded
from a match.